fix: moved the wrapper function to the help in pkg/views/util
